ai-service/requirements.txt
flask
flask-cors
transformers
torch
pypdf
python-docx
language-tool-python
python-dotenv
groq
